cli/requirements.txt
deepface==0.0.93
tf-keras
numpy==1.26.4
scikit-learn==1.4.2
requests-oauthlib==1.3.1
requests==2.31.0
httpx[http2]==0.27.0
Pillow==10.3.0
supabase==2.4.0
python-dotenv==1.0.0
